Refactor subsidy module
