+++ zebra-consensus/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `block::subsidy::funding_stream_address`
+
+### Removed
+
+- `block::subsidy::new_coinbase_script`
+- `block::subsidy::funding_streams`
+
 ## [3.1.0] - 2025-11-17
 
 ### Added
