+++ zebra-consensus/src/lib.rs
@@ -47,10 +47,7 @@ pub mod transaction;
 #[cfg(any(test, feature = "proptest-impl"))]
 pub use block::check::difficulty_is_valid;
 
-pub use block::{
-    subsidy::{funding_streams::funding_stream_address, new_coinbase_script},
-    Request, VerifyBlockError, MAX_BLOCK_SIGOPS,
-};
+pub use block::{subsidy::funding_stream_address, Request, VerifyBlockError, MAX_BLOCK_SIGOPS};
 pub use checkpoint::{VerifyCheckpointError, MAX_CHECKPOINT_BYTE_COUNT, MAX_CHECKPOINT_HEIGHT_GAP};
 pub use config::Config;
 pub use error::BlockError;
